sports_science_dataset/requirements.txt
# Include shared dependencies
-r ../requirements.txt

# Database and ORM
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1

# Scientific APIs and data collection
biopython==1.83
aiohttp==3.9.1
httpx==0.25.2

# PDF processing and text extraction
PyMuPDF==1.23.14
pdfplumber==0.10.3
python-magic==0.4.27

# NLP and embeddings
sentence-transformers==2.2.2
transformers==4.36.2
torch==2.1.2
numpy==1.24.4
scikit-learn==1.3.2

# OpenAI API for AI filtering
openai==1.6.1

# Data processing and utilities
pandas==2.1.4
tqdm==4.66.1
pydantic==2.5.2
pydantic-settings==2.1.0

# Logging and monitoring
loguru==0.7.2
rich==13.7.0

# Rate limiting and retries
tenacity==8.2.3
ratelimit==2.2.1

# File handling and utilities
pathlib2==2.3.7.post1
click==8.1.7
pyyaml==6.0.1

# Testing and development
pytest-asyncio==0.21.1
flake8==7.0.0
mypy==1.8.0

# Additional utilities
validators==0.22.0
fuzzywuzzy==0.18.0
python-levenshtein==0.23.0